requirements.txt
flask
redis
requests
openai
python-dotenv